Add `useCachedMessages` hook, update `useMessages`
This PR adds message caching via IndexedDB in a browser when using the new
`useCachedMessages` hook. The new hook has the same API as the `useMessages`
hook, but will add messages to a cache and attempt to fetch messages from the
cache before accessing the network.

It also optimizes the existing `useMessages` hook to match some of the new
logic in the `useCachedMessages` hook.